requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
# antlr4-python3-runtime==4.9.3
# anyio==3.6.2
# aspose-pdf==23.5.0
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==39.0.2
cv==1.0.0
cycler==0.11.0
Cython==0.29.35
dataclasses==0.6
DateTime==5.1
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.10.0
filetype==1.2.0
fire==0.5.0
Flask==2.2.3
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
fvcore==0.1.5.post20221221
gevent==22.10.2
gradio-client==0.2.6
gradio==3.34.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
ImageHash==4.3.1
imageio==2.26.1
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
labelme2coco==0.2.4
lazy-loader==0.1
linkify-it-py==2.0.2
llvmlite==0.40.1rc1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.57.0
numpy==1.24.2
omegaconf==2.3.0
onnxruntime==1.15.0
opencv-python-headless==4.7.0.72
opencv-python==4.7.0.72
orjson==3.9.1
outcome==1.2.0
packaging==23.0
pandas==2.0.2
pathspec==0.11.1
pdf2image==1.16.3
Pillow==9.4.0
pip==22.3.1
pixellib==0.7.1
platformdirs==3.5.3
plyer==2.1.0
pooch==1.7.0
portalocker==2.7.0
postmarker==1.0
protobuf==4.23.2
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.6
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.8
pymongo==4.3.3
pyparsing==3.0.9
# PyQt5-Qt5==5.15.2
# PyQt5-sip==12.12.1
# PyQt5==5.15.9
pyreadline3==3.4.1
pyrsistent==0.19.3
PySimpleGUI==4.60.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
# pywin32==306
PyYAML==6.0
rembg==2.0.43
requests==2.28.2
sahi==0.11.14
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.3
semantic-version==2.10.0
setuptools==65.5.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.26.1
suncalc==0.1.2
sympy==1.11.1
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.3.15
toolz==0.12.0
torch==2.0.0
tqdm==4.65.0
trio-websocket==0.10.2
trio==0.22.0
typing-extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.14
uvicorn==0.21.1
watchdog==3.0.0
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.2.3
wsproto==1.2.0
xmljson==0.2.1
yacs==0.1.8
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
